+++ src/DefaultNode.php
@@ -29,7 +29,11 @@ class DefaultNode implements Node
 
     public function getParentNodeId(): ?string
     {
-        return (string) $this->getNodeEntry('parent_id');
+        if($parentId = $this->getNodeEntry('parent_id')) {
+            return (string) $parentId;
+        }
+
+        return null;
     }
 
     public function equalsNode(Node $other): bool